+++ src/refactor_nrepl/ns/suggest_aliases.clj
@@ -1,6 +1,7 @@
 (ns refactor-nrepl.ns.suggest-aliases
   "Suggestion of aliases based on these guidelines: https://stuartsierra.com/2015/05/10/clojure-namespace-aliases"
   (:require
+   [clojure.edn :as edn]
    [clojure.string :as string]))
 
 (defn test-like-ns-name? [ns-sym]
@@ -11,6 +12,14 @@
                    (or (string/starts-with? last-fragment "t-")
                        (string/starts-with? last-fragment "test-")
                        (some #{"test" "unit" "integration" "acceptance" "functional" "generative"} fragments)))))))
+
+(def readable-as-symbol?
+  (memoize
+   (fn [s]
+     (try
+       (symbol? (edn/read-string s))
+       (catch Exception _
+         false)))))
 
 (defn suggested-aliases [namespace-name]
   (let [fragments (-> namespace-name str (string/split #"\."))
@@ -29,8 +38,10 @@
                        (range 1 (inc (count fragments)))
                        (repeat (distinct fragments)))
         v (into {}
-                (map (fn [segments]
-                       [(->> segments (string/join ".") (symbol)),
-                        [namespace-name]]))
+                (keep (fn [segments]
+                        (let [candidate (->> segments (string/join "."))]
+                          (when (readable-as-symbol? candidate)
+                            [(symbol candidate),
+                             [namespace-name]]))))
                 fragments)]
     (dissoc v namespace-name)))